third_party/WebKit/LayoutTests/webexposed/web-animations-api.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
    assert_true('animate' in Element.prototype);
}, 'Element.animate() should be exposed.');

test(function() {
    assert_false('Animation' in window);
},'Animation constructor should not be exposed without experimental web platform features enabled.\nThis test is expected to fail in LayouTests/webexposed.');

test(function() {
    assert_false('Timeline' in window);
},'Timeline should not be exposed without experimental web platform features enabled.\nThis test is expected to fail in LayouTests/webexposed.');

test(function() {
    assert_false('timeline' in document);
},'document.timeline should not be exposed without experimental web platform features enabled.\nThis test is expected to fail in LayouTests/webexposed.');

test(function() {
    assert_false('Timing' in window);
},'Timing should not be exposed without experimental web platform features enabled.\nThis test is expected to fail in LayouTests/webexposed.');

test(function() {
    assert_false('TimedItem' in window);
},'TimedItem should not be exposed without experimental web platform features enabled.\nThis test is expected to fail in LayouTests/webexposed.');
</script>
